+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   rename `replaceIndividualInputAttributes()` to `individualInputAttributes()` and `replaceRadioAttributes()` 
   to `radioAttributes()` (@vjik)
 - Enh #142: Make `NoEncodeStringableInterface` extend from `Stringable` interface (@vjik)
+- Bug #143: Fix a typo in the message of exception that thrown on invalid buttons' data in `ButtonGroup` widget (@vjik)
+- Enh #143: Minor type hinting improvements (@vjik)
 
 ## 2.5.0 July 09, 2022